Add tests for authtoken utilization. These tests run over the course of an auth token's timeout, and take a long time (20 min each). (#24)
They are in seperate files to enable jest to run them in parallel and reduce the amount of clock time spent.

Also moved existing reconnect tests to seperate files for parallelism.

Instead of adding long running tests with .skip tagging in the code, there's now a setting for this so they can be controlled at runtime. Currently off by default.

Also, found that the message for the auth token expiring and not refreshing cleanly was being eaten and some bad promise chain being invoked instead of the error being cleanly returned. Fixed.