Add proxy support for nodeJS. (#40)
As the proxy package depends on node features, when the webpack is built this change will skip building it in and will replace any use of it with an exception. Since the use is conditional to using a WebSocket library that also only works in node, this should be safe.