Add OCSP certificate validation when connecting from node.js, including via proxy.
Imported the OCSP library from https://github.com/indutny/ocsp with a modification to allow the OCSP validation to work via proxy. (Will be working out of band to get that change contributed back and then revert to consuming directly instead.)

Changed the gulpfile to use the tsconfig.json in the root directory instead of having a seperate copy of the settings.